+++ test-parts/part-i-api_frontend/front-main.js
@@ -5,10 +5,8 @@ const cCheckListFile = require("./parts/d-check_device_list");
 const cListAvaliableFile = require("./parts/e-list_avaliable_nodes");
 const cCheckNodeArrayFile = require("./parts/f-check_node_array");
 const cGetDeviceStatusFile = require("./parts/g-get_device_properties");
-/*
 const cDeleteDevicesFile = require("./parts/h-delete_added_devices");
-const cClearCacheFile = require("./parts/i-clear_cache");
-*/
+//const cClearCacheFile = require("./parts/i-clear_cache");
 
 
 function coordinateFrontendTesting()
@@ -100,7 +98,7 @@ function callGetDeviceStatus()
 		
 		after(function()
 		{
-			//callDeleteDevices();
+			callDeleteDevices();
 		});
 		
 	});
@@ -114,7 +112,7 @@ function callDeleteDevices()
 		
 		after(function()
 		{
-			callClearCache();
+			//callClearCache();
 		});
 		
 	});